Refactor Flask app into package structure
