+++ zk_tools_web/static/css/styles.css
@@ -1,0 +1,12 @@
+body { background-color: #f8f9fa; }
+main { padding-bottom: 4rem; }
+.selected-row { background-color: #e8f4ff !important; }
+.biometric-list { margin-bottom: 0; padding-left: 1.25rem; }
+.accordion-button:not(.collapsed) { background-color: #0d6efd; color: #fff; }
+.accordion-button:focus { box-shadow: none; }
+.dataTables_wrapper .dataTables_paginate .paginate_button { border-radius: 0.375rem; }
+.dataTables_wrapper .dataTables_paginate .paginate_button.current {
+    background-color: #0d6efd !important;
+    color: #fff !important;
+}
+.dataTables_wrapper .dataTables_length select { width: auto; }
